Update to Habbo Origins v142 crypto

+++ G-Earth/src/main/java/gearth/protocol/crypto/RC4Cipher.java
@@ -10,7 +10,7 @@ public interface RC4Cipher {
 
     byte[] decipher(byte[] data, int offset, int length);
 
-    byte[] getState();
+    int[] getState();
 
     int getQ();
